+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/Ibea/IbeaReferences.cs
@@ -1,0 +1,10 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.Ibea;
+public static class IbeaReferences
+{
+    public static ScientificReference ZitzlerKunzli2004 { get; } =
+        new("Zitzler & Kunzli", 2004,
+            "Indicator-Based Selection in Multiobjective Search",
+            "PPSN VIII, LNCS 3242, 832-842",
+            "10.1007/978-3-540-30217-9_84");
+}